# Added ability to de/serialize from/to avro container.