[Avro Container](https://avro.apache.org/docs/1.11.1/specification/#object-container-files)

* Added `AvroConvert.DeserializeContainer()` for deserialization (of IEnumerable<>).
* Added `AvroConvert.SerializeContainer()` for serialization (of IEnumerable<>).
* Added `IsEnumerable()` to `Type` extensions.
* Workaround for deserialization added to `Resolver.ResolveArray()`
* Added some [basic] tests.

*** I haven't had time to test everything as I only needed serialization for Azure Data Explorer