@wdcossey 2023-10-05
> Thank you for your contribution, I will do my best to review it today. Looks pretty nice at first glance!

Pushed a bug-fix and some enhancements.