@AdrianStrugala 2024-03-03
Hey,
I am going to implement this in a similar way that you've suggested Manvel.
The point is, that this is in fact a breaking change and I would make it part of the v4 release.
Adrian